Add selectors for Clojure/ClojureScript.
The scope for ClojureScript should be changing to source.clojure.clojurescript
(as in https://github.com/sublimehq/Packages/pull/2355) but others will
continue to use the older scope. This update makes that work.

The name of the key that users can use to add docsets is `docsets_user`.
The documentation didn't reflect that.